Moving Maintenance Profile to TypeScript for better types
