+++ src/routes/features/maintenance/[id]/+page.svelte
@@ -7,7 +7,7 @@
     import ProfileView from "$components/maintenance/ProfileView.svelte";
 
     const profileId = $page.params.id;
-    /** @type {import('$lib/extension/entities/MaintenanceProfile.js').default|null} */
+    /** @type {import('$entities/MaintenanceProfile.ts').default|null} */
     let profile = null;
     let isActiveProfile = false;
